Fix rounding error in roundrect, fix hittest for hardware triangles
git-svn-id: https://nekonme.googlecode.com/svn/trunk@1439 1509560c-5e2a-0410-865c-31c25e1cfdef

+++ project/common/Graphics.cpp
@@ -95,13 +95,13 @@ void Graphics::drawRoundRect(float x,float  y,float  width,float  height,float  
    float w = width*0.5;
    x+=w;
    if (rx>w) rx = w;
-   int   lw = w - rx;
+   float lw = w - rx;
    float w_ = lw + rx*SIN45;
    float cw_ = lw + rx*TAN22;
    float h = height*0.5;
    y+=h;
    if (ry>h) ry = h;
-   int   lh = h - ry;
+   float lh = h - ry;
    float h_ = lh + ry*SIN45;
    float ch_ = lh + ry*TAN22;
